Refactor supercontract mongo extension & bug fixes

+++ extensions/mongo/plugins/supercontract_v2/src/AutomaticExecutionsPaymentMapper.cpp
@@ -16,7 +16,7 @@ namespace catapult { namespace mongo { namespace plugins {
     template<typename TTransaction>
     void StreamAutomaticExecutionsPaymentTransaction(bson_stream::document& builder, const TTransaction& transaction) {
         builder << "contractKey" << ToBinary(transaction.ContractKey)
-                << "automaticExecutionsNumber" << ToInt64(transaction.AutomaticExecutionsNumber)
+                << "automaticExecutionsNumber" << ToInt32(transaction.AutomaticExecutionsNumber)
                 << "signer" << ToBinary(transaction.Signer);
     }
